# api/openapi.yaml
components:
  schemas:
...
    ApiHTTPError:
      additionalProperties: true
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
      type: object
...
    CLIAddCalendarPlanRequest:
      additionalProperties: false
      properties:
        email:
          type: string
        headless:
          type: boolean
        oauth_app:
          type: string
        oauth_app_explicit:
          type: boolean
      required:
        - email
      type: object
    CLIAddCalendarPlanResponse:
      additionalProperties: true
      properties:
        body_lines:
          items:
            type: string
          type:
            - array
            - "null"
        cancel_hint:
          type: string
        headline:
          type: string
        needs_client_check:
          type: boolean
        needs_scope_escalation:
          type: boolean
        oauth_app:
          type: string
        oauth_app_resolved:
          type: boolean
      required:
        - needs_scope_escalation
      type: object
...
    CLIDeduplicatePlanItem:
      additionalProperties: true
      properties:
        backfilled_count:
          format: int64
          type: integer
        duplicate_messages:
          format: int64
          type: integer
        needs_confirmation:
          type: boolean
        plan_fingerprint:
          type: string
        scope_is_collection:
          type: boolean
        scope_label:
          type: string
        source_id:
          format: int64
          type: integer
        stdout:
          type: string
      required:
        - needs_confirmation
      type: object
    CLIDeduplicatePlanRequest:
      additionalProperties: false
      properties:
        account:
          type: string
        collection:
          type: string
        content_hash:
          type: boolean
        delete_dups_from_source_server:
          type: boolean
        prefer:
          type: string
      type: object
    CLIDeduplicatePlanResponse:
      additionalProperties: true
      properties:
        footer_stdout:
          type: string
        items:
          items:
            $ref: "#/components/schemas/CLIDeduplicatePlanItem"
          type:
            - array
            - "null"
        prefix_stdout:
          type: string
      required:
        - items
      type: object
    CLIDeleteStagedPlanRequest:
      additionalProperties: false
      properties:
        account:
          type: string
        batch_id:
          type: string
        dry_run:
          type: boolean
        list:
          type: boolean
        permanent:
          type: boolean
        remote_delete_enabled:
          type: boolean
        "yes":
          type: boolean
      type: object
    CLIDeleteStagedPlanResponse:
      additionalProperties: true
      properties:
        blocked_error:
          type: string
        confirmation_mode:
          type: string
        needs_confirmation:
          type: boolean
        needs_execution:
          type: boolean
        needs_scope_escalation:
          type: boolean
        plan_fingerprint:
          type: string
        planned_batch_ids:
          items:
            type: string
          type:
            - array
            - "null"
        remote_delete_env_var:
          type: string
        scope_escalation_account:
          type: string
        scope_escalation_body_lines:
          items:
            type: string
          type:
            - array
            - "null"
        scope_escalation_cancel_hint:
          type: string
        scope_escalation_headline:
          type: string
        scope_escalation_oauth_app:
          type: string
        stdout:
          type: string
      required:
        - needs_execution
        - needs_confirmation
      type: object
...
    CLIEmbeddingsPlanRequest:
      additionalProperties: false
      properties:
        force:
          type: boolean
        generation_id:
          format: int64
          type: integer
        operation:
          type: string
      required:
        - operation
        - generation_id
      type: object
    CLIEmbeddingsPlanResponse:
      additionalProperties: true
      properties:
        needs_confirmation:
          type: boolean
        prompt:
          type: string
      required:
        - needs_confirmation
      type: object
...
    CLIRunEvent:
      additionalProperties: true
      properties:
        data:
          type: string
        error:
          type: string
        type:
          type: string
      required:
        - type
      type: object
    CLIRunRequest:
      additionalProperties: false
      properties:
        args:
          items:
            type: string
          type:
            - array
            - "null"
        cwd:
          type: string
        env:
          additionalProperties:
            type: string
          type: object
      required:
        - args
      type: object
...
    ErrorResponse:
      additionalProperties: true
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
      type: object
...
  securitySchemes:
    apiKey:
      in: header
      name: X-Api-Key
      type: apiKey
info:
  title: msgvault API
  version: 1.24.0
openapi: 3.1.0
paths:
...
  /api/v1/cli/add-calendar/plan:
    post:
      operationId: planCLIAddCalendar
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CLIAddCalendarPlanRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CLIAddCalendarPlanResponse"
          description: OK
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Error
      security:
        - apiKey: []
      summary: Plan CLI Calendar account setup
      tags:
        - API
...
  /api/v1/cli/deduplicate/plan:
    post:
      operationId: planCLIDeduplicate
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CLIDeduplicatePlanRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CLIDeduplicatePlanResponse"
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiHTTPError"
          description: Bad Request
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiHTTPError"
          description: Unprocessable Entity
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiHTTPError"
          description: Internal Server Error
        "503":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiHTTPError"
          description: Service Unavailable
      security:
        - apiKey: []
      summary: Plan interactive deduplication for CLI use
      tags:
        - CLI
...
  /api/v1/cli/delete-staged/plan:
    post:
      operationId: planCLIDeleteStaged
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CLIDeleteStagedPlanRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CLIDeleteStagedPlanResponse"
          description: OK
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Error
      security:
        - apiKey: []
      summary: Plan CLI staged deletion execution
      tags:
        - API
...
  /api/v1/cli/embeddings/plan:
    post:
      operationId: planCLIEmbeddings
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CLIEmbeddingsPlanRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CLIEmbeddingsPlanResponse"
          description: OK
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Error
      security:
        - apiKey: []
      summary: Plan CLI embeddings management
      tags:
        - API
...
  /api/v1/cli/run:
    post:
      operationId: runCLI
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CLIRunRequest"
        required: true
      responses:
        "200":
          content:
            application/x-ndjson:
              schema:
                $ref: "#/components/schemas/CLIRunEvent"
          description: OK
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Error
      security:
        - apiKey: []
      summary: Run an allowlisted CLI command
      tags:
        - API
...